Alternate strategy,to just pass back responses.
