+++ speech/google/cloud/speech/__init__.py
@@ -16,3 +16,4 @@
 
 from google.cloud.speech.client import Client
 from google.cloud.speech.connection import Connection
+from google.cloud.speech.encoding import Encoding
