+++ speech/google/cloud/speech/streaming_result.py
